A great rework awaits...
This right now is not complete, still needs to merge with dev to get the restructuring, and I have a final few changes to make. Just getting this off of my computer for the time being, and somewhere a bit safer.

Overall this massive rework includes:
-Completely new armor system that is volumetric
-Primitive prop integration (replaces pArmor on spawn)
-Long awaited HE rework (not quite finished, but nearly there)
-New health calculations (got rid of the godforsaken ACF.Threshold value)

Still to be done:
-Finish HE rework (working on a new fragment method)
-New file loader integration
-Whatever else pops up in development hell

This commit alone is by no means ready to go to dev